Some Refactoring
- GPXTableView
- move things from GPXEditor to tableview classes
